Refactor type checking functions for integers and strings; remove debug output from getValue and filter tests

+++ root/source/uim/root/datatypes/jsons/types/integers/typecheck.d
@@ -138,7 +138,7 @@ unittest {
 // #endregion index
 
 bool isInteger(Json json) {
-  return (json.type == Json.Type.float_);
+  return (json.type == Json.Type.int_);
 }
 ///
 unittest {
